change in appointments

+++ src/main/java/rs/ac/bg/fon/chatbot/db/services/repositories/OfficeHoursRepository.java
@@ -6,7 +6,7 @@ import rs.ac.bg.fon.chatbot.db.domain.OfficeHours;
 
 import java.util.List;
 
-public interface OfficeHoursRepository extends CrudRepository<OfficeHours, Integer> {
+public interface OfficeHoursRepository extends CrudRepository<OfficeHours, Long> {
 
     @Query("select o from OfficeHours o, Professor p where o.professor = p and p.email = ?1 order by o.beginTime desc")
     List<OfficeHours> findAllByProfessorEmail(String email);
